Differentiate links from image type and stylesheet
Uses endswith() method rather than using type.  When 'ng build' renders the index.html for example it wont' always put ['type'] in <link>.  So this way we know to look for .css or if an image, we look for e.g, .ico.  Most likely were going to have file-types in different directories within static.